Extend the forest search with queries.
A set of queries may now be passed to forest_search.
For each element that matches the search callback,
a second query callback is called with the active queries.
The search only continues if this callback is true for at least
one query.
The queries for which the query callback returns false are
deactivated for the remaining recursion (the children are not
called for these queries).

If no queries are given the query callback is not called and
the search callback alone decides whether the recursion continues
or not.