Add a query index to the search callback.
The search callback gets as additional information
the index of the query in the global queries array.
This allows us easily to identify the current query.
An example usage would be to pass an array of lenght
number of queries as user pointer and store the matching
elements for each query and the correct position in
this array.