Correct the point inside check for lines and triangles.
Checking whether the point is inside each plane defined
by every face does only work in 3D.
For 2D triangles we solve a system of linear equations,
testing whether we can write the point as linear combination
of the triangle's vertices.
For lines we solve a similar equation with onlt one vector.
For points (T8_ECLASS_VERTEX) we just check whether
the point to check equals the vertex (up to a tolerance).

There is no implementation for quads yet.