Fix long/lat to xyz computation.
Fix a bug in the computation of euclidean
coordinates from longitude and latitude.
First, we assumed that long and lat were
given in radians but they are actually given
in degrees, so we added a conversion.
Second, latitude ranges from -90 to 90,
hence we need to switch sin and cos compared
to 0 to 180.