Change install dirs to include subdirs
Before: Every header was installed into
  INSTALL/include
Now: The header are installed into
  INSTALL/include/src/
with subfolders.
Thus the t8_data/t8_shmem.h header for example is
installed to
  INSTALL/include/src/t8_data

We would like to get rid of the additional src/ in the
include path in the future.
Thus, install to
  INSTALL/include/

However, we did not figure it out yet without using
SUBDIRS and recursive Make