Deactivate T8_ASSERT in non-debug mode.
This is necessary since we may link t8code with a p4est in
debug mode, but want t8code without debug mode (exotic case
but it should be supported).
In this case the assertion was executed which resulted in compiler
errors
when t8code variables where asserted that only lived in debugging mode.