Add t8_forest_set_subelements (eliminate_hanging_nodes) and t8_forest_subelements as extra features next to adapt, balance etc.
There is still work to do. We need to make these new functions compatible with from_method and set_from.
At the moment, we are able to refine all elements with subelements, after we adapted the mesh (without balance).
We should be able to do this after balance and we should be able to refine only some elements by a specific type of subelement.