Define the specific callback function for the refine values in the subelement scheme
The callback function is not working at the moment. There is still a problem, using set_subelements
after set_balance. Since some functions in the callback rely on balanced meshes, we can not use them at the moment.
The main question is how to properly implement set_subelements into the high level infrastructure of t8.
If we can use the now defined callback funtion, plotting different subelements at the right spots of a balanced mesh
should not be a problem.