Add the feature, that we are now able to insert subelements into a balanced mesh, at the right spots (if the level of at least one neighbor is higher)
There are some memory issues left, but at least we can plot the results, which are looking good. We are also able to determine different
types of subelements, depending on which neighbors have a higher refinement level. This can be used to insert the right subelement types
to eliminate hanging nodes.