Fix a big where recursive adaptation was falsely entered.
The recursion should only be entered for the last child of
a family.
This meant that it was also enteres for elements that only
have one child (i.e. vertices in the default scheme).
However, for this child the adapt callback was already called,
so recursively calling it is not necessary.

Co-authored-by: David Knapp <david.knapp@dlr.de>