Implement a new element function to determine the number of subelements, based on the subelement_type
We need this function because we need to allocate enough memory in t8_forest_adapt before building the subelements.
Next task: implement a element_get_location_of_subelement function which will help to automatically build the different
subelements and their vertecies, given the type and id of the whole element.