Correct some subelement functions and make them easier.
There is a memory problem, if we change from the temporary subelements to the real ones. It seems like we allocate memory
for four children in forest_adapt, which works for two subelements but not for types with more than 4 subelements. Maybe,
this issue is automatically fixed with the merge.