Extend the subelement example
Our aim is to adapt a mesh with subelements several times.
The idea is to implement a new callback function, that returns -1 if an element is a subelement and 0 else.
Set_adapt checks, whether there are subelements in the given forest. If so, the forest gets adapted with the new callback function
such that no subelements are left in the forest before adapting it using the user data.