Continue outsource operations to element level
Added the t8_element function find_neighbor_in_transition_cell. The idea is the following. We only need to solve the
case that our neighbor element is a subelement, since in this case we will get a random subelement of the right transition cell
but which might be no neighbor. In the new function we will identify the right subelement and output its subelement_id. This id
can be used to adjust the element_index and then get the right neighbor element from the element_array.

Also deleted an unnecessary function.