Remove the adapt_static scheme from the advect_with_subelements example and use the adapt-freq argument for it
Corresponding adjustements have been made. We are now able to solve the linear advection for arbitrary initial static meshes.
The following task is to be able to solve the equation with an adaptive mesh. At the moment there is some bug such that it does not
even work without subelements. Further investigations have to be made here. It seems that if an element is refined, no flux will be computed
into this family anymore.