Introduce new shmem_init and finalize functions
Before we can use shared memory the inter- and
intronode communicators need to be attached to
an MPI comm.
We added new t8_shmem functions to attach and
detach them.
The init function must be called for a comm before
using shared memory.
We ensured to call it in t8code in each situation (currently
cmesh_partition only) where shared memory is used.