Fix wrong argument forwarding in t8_cmesh_set_attribute/t8_stash_add_attribute
t8_cmesh_set_attribute expects a flag 'data_persists' that if true
specifies that the attribute data will still be present at commit time.
Internally this is passed to t8_stash_add_attribute which expects a
flag 'copy', doing the same but with inverted truth values.
The data_persist flag was passed through without negating it,
resulting in it being wrong.