Pyra-Corner Bugfix and Volumecomputation for pyramids
Found a bug in the computation of the number of corners for pyramids.
It was fixed to 5, but a pyra in the shape of a tet has 4 corners. Fixed it.

Added the computation of the volume of a pyramid by dividing it into two tets and to sum
over the volume of the tets.