Use a workaround in LFN such that it also works for multiple processes and ghost neighbor elements.
But there is still work to do: At this momment it is not clear whether the right subelement neighbor is identified,
if it is a ghost element (since we do not adjust it at the moment). We can probably not use the standard procedure
as it iterates over all subelements of a transition cell but in a ghost structure there might only be some subelements
of a family of subelements.