Triangle Bugfix: CHECK THIS WITH EXTRA CARE
The id of a tri/tet (and therefore prisms too) was computed wrong,
if the level in the input-argument was set lower than t->level.
Before entering the for-loop the level was always set to t->level, hence the id was always
computed on t->level. This push fixes this problem. If the level is larger than t->level the type
is set manually (the type of the descendants with the same id is the same as the type of t).
Afterwards the level is set to my_level.
Otherwise we do not precompute the exponent and the temporary type is computed in the old fashion,
but on level level, instead of level t->level.