Revert all ubuntu versions of github action to 20.04
Why? Before it was ubuntu-latest.
On December 1 github switched from Ubuntu 20.04 to 22.04.
This broke our tests, since the test script does not compile under the
new ubuntu (note that this does not mean that t8code does not run on
22.04).
Until the issue is resolved we revert manually back to 22.04.